requirements.txt
amqp==5.0.5
appdirs==1.4.4
asgiref==3.3.1
autopep8==1.5.5
billiard==3.6.3.0
cairocffi==1.2.0
CairoSVG==2.5.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
confusable-homoglyphs==3.2.0
cssselect2==0.4.1
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.5
django-admin-interface==0.15.3
django-autoslug==1.9.8
django-colorfield==0.4.1
django-countries==7.0
django-crispy-forms==1.11.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-grappelli==2.14.3
django-heroku==0.3.1
django-registration==3.1.1
django-tastypie==0.14.3
filelock==3.0.12
gunicorn==20.0.4
html5lib==1.1
kombu==5.0.2
mysqlclient==2.0.3

olefile==0.46
pep8==1.7.1
Pillow==8.1.0
pipenv==2020.11.15
prompt-toolkit==3.0.16
psycopg2==2.8.6
pycairo==1.20.0
pycodestyle==2.6.0
pycparser==2.20
Pyphen==0.10.0

python-dateutil==2.8.1
python-decouple==3.4
python-mimeparse==1.6.0
pytz==2019.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.0
tinycss2==1.1.0
toml==0.10.2
Unipath==1.1
vine==5.0.0
virtualenv==20.4.2
virtualenv-clone==0.5.4
wcwidth==0.2.5
WeasyPrint==52.2
webencodings==0.5.1
whitenoise==5.2.0
